requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0

# Core Data
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.3.0

# Deep Learning / NLP
torch==2.0.1
transformers==4.38.2
sentence-transformers==2.6.1
huggingface-hub==0.21.4
sentencepiece==0.2.1
sacremoses==0.1.1

# Vector Database
# faiss-cpu==1.8.0

# Google AI APIs
google-generativeai==0.3.0
google-cloud-vision==3.4.1

# Image Processing
pillow==10.0.0

# Web + Parsing
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Language Processing
langdetect==1.0.9
textblob==0.17.1
vaderSentiment==3.3.2
spacy==3.6.0
deep-translator==1.11.4

# Scientific Papers
arxiv==1.4.7

# Date Handling
python-dateutil==2.8.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Validation
pydantic==1.10.13